+++ modules/src/ics07_tendermint/error.rs
@@ -12,7 +12,7 @@ pub enum Kind {
     #[error("invalid trusting period")]
     InvalidTrustingPeriod,
 
-    #[error("invalid client state trust threshold ")]
+    #[error("invalid client state trust threshold")]
     InvalidTrustThreshold,
 
     #[error("invalid unbonding period")]