+++ relayer-cli/src/commands/listen.rs
@@ -3,6 +3,7 @@ use std::{fmt, ops::Deref, str::FromStr, sync::Arc, thread};
 use abscissa_core::{application::fatal_error, error::BoxError, Command, Options, Runnable};
 use itertools::Itertools;
 use tokio::runtime::Runtime as TokioRuntime;
+use tracing::{error, info};
 
 use ibc::{events::IbcEvent, ics24_host::identifier::ChainId};
 
@@ -92,8 +93,8 @@ impl Runnable for ListenCmd {
 
 /// Listen to events
 pub fn listen(config: &ChainConfig, filters: &[EventFilter]) -> Result<(), BoxError> {
-    println!(
-        "[info] Listening for events `{}` on '{}'...",
+    info!(
+        "listening for events `{}` on '{}'...",
         filters.iter().format(", "),
         config.id
     );
@@ -116,15 +117,15 @@ pub fn listen(config: &ChainConfig, filters: &[EventFilter]) -> Result<(), BoxEr
                     continue;
                 }
 
-                println!("- Event batch at height {}", batch.height);
+                info!("- event batch at height {}", batch.height);
 
                 for event in matching_events {
-                    println!("+ {:#?}", event);
+                    info!("+ {:#?}", event);
                 }
 
-                println!();
+                info!("");
             }
-            Err(e) => println!("- Error: {}", e),
+            Err(e) => error!("- error: {}", e),
         }
     }