+++ relayer-cli/src/commands/upgrade.rs
@@ -2,7 +2,7 @@
 
 use abscissa_core::{Command, Help, Options, Runnable};
 
-use crate::commands::tx::client::{TxUpgradeClientCmd, TxUpgradeClientsCmd};
+use crate::commands::tx::client::TxUpgradeClientCmd;
 
 #[derive(Command, Debug, Options, Runnable)]
 pub enum UpgradeCmds {
@@ -13,8 +13,10 @@ pub enum UpgradeCmds {
     /// Subcommand for upgrading a `client`
     #[options(help = "Upgrade an IBC client")]
     Client(TxUpgradeClientCmd),
-
-    /// Subcommand for upgrading all `client`s that target specified chain
-    #[options(help = "Upgrade all IBC clients that target a specific chain")]
-    Clients(TxUpgradeClientsCmd),
+    //
+    // NOTE: This command is disabled until https://github.com/informalsystems/ibc-rs/issues/1229 is fixed.
+    //
+    // /// Subcommand for upgrading all `client`s that target specified chain
+    // #[options(help = "Upgrade all IBC clients that target a specific chain")]
+    // Clients(TxUpgradeClientsCmd),
 }