+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-## UNRELEASED
+## Unreleased
+
+> Nothing yet
+
+## v0.6.1
+*July 22nd, 2021*
+
+This minor release mainly improves the reliability of the relayer
+by ensuring that pending packets are cleared on start,
+and that Hermes can recover from the WebSocket subscriptions
+being closed under its feet by Tendermint.
+
+Upgrading from version `0.6.0` to `0.6.1` requires no explicit steps.
+
+> **WARNING:** Due to a regression ([#1229]), the `upgrade client`,
+> `tx raw upgrade-clients`, and `tx raw upgrade-chain` commands have
+> been temporarily disabled in this version.
+> These commands will be re-enabled in the next version.
 
 ### FEATURES
 
@@ -8,9 +25,6 @@
   - Enable `pub` access to verification methods of ICS 03 & 04 ([#1198])
   - Add `ics26_routing::handler::decode` function ([#1194])
   - Add a pseudo root to `MockConsensusState` ([#1215])
-
-- [ibc-relayer-cli]
-  - Added `upgrade-clients` CLI ([#763])
 
 ### IMPROVEMENTS
 
@@ -25,14 +39,19 @@
   - Align `as_str` and `from_str` behavior in `ClientType` ([#1192])
 
 - [ibc-relayer]
-  - Fixed: Hermes does not clear packets on start ([#1200])
+  - Ensure pending packets are cleared on start ([#1200])
+  - Recover from missed RPC events after WebSocket subscription is closed by Tendermint ([#1196])
+
 
 [#1094]: https://github.com/informalsystems/ibc-rs/issues/1094
 [#1114]: https://github.com/informalsystems/ibc-rs/issues/1114
 [#1192]: https://github.com/informalsystems/ibc-rs/issues/1192
 [#1194]: https://github.com/informalsystems/ibc-rs/issues/1194
+[#1196]: https://github.com/informalsystems/ibc-rs/issues/1196
 [#1198]: https://github.com/informalsystems/ibc-rs/issues/1198
 [#1200]: https://github.com/informalsystems/ibc-rs/issues/1200
+[#1215]: https://github.com/informalsystems/ibc-rs/issues/1215
+[#1229]: https://github.com/informalsystems/ibc-rs/issues/1229
 
 
 ## v0.6.0
@@ -119,7 +138,6 @@ The full list of changes is described below.
 [#69]: https://github.com/informalsystems/ibc-rs/issues/69
 [#600]: https://github.com/informalsystems/ibc-rs/issues/600
 [#697]: https://github.com/informalsystems/ibc-rs/issues/697
-[#763]: https://github.com/informalsystems/ibc-rs/issues/763
 [#1062]: https://github.com/informalsystems/ibc-rs/issues/1062
 [#1117]: https://github.com/informalsystems/ibc-rs/issues/1117
 [#1057]: https://github.com/informalsystems/ibc-rs/issues/1057